# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * GPay Market to Digiseller synchronization API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  invoiceId: string;
  digisellerProductId: number;
  productName: string;
  /** @nullable */
  paidAmountRub: number | null;
  saleTimestamp: Date;
  status: OrderStatus;
  /** @nullable */
  operatorNote: string | null;
  syncedAt: Date;
  updatedAt: Date;
}