# lib/api-zod/src/generated/types/orderPage.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * GPay Market to Digiseller synchronization API
 * OpenAPI spec version: 0.1.0
 */
import type { Order } from './order';

export interface OrderPage {
  items: Order[];
  total: number;
  page: number;
  pageSize: number;
}