# lib/api-zod/src/generated/types/orderSyncResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * GPay Market to Digiseller synchronization API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderSyncResult {
  fetched: number;
  inserted: number;
  updated: number;
  ignored: number;
  skipped: boolean;
}