# lib/api-zod/src/generated/types/orderUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * GPay Market to Digiseller synchronization API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderUpdateStatus } from './orderUpdateStatus';

export interface OrderUpdate {
  status?: OrderUpdateStatus;
  /**
     * @maxLength 2000
     * @nullable
     */
  note?: string | null;
}